auth/requirements.txt
base==0.0.0
blinker==1.8.2
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
flask==3.0.3
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
PyJWT==2.10.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
pycparser==2.22
SQLAlchemy==2.0.37
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
gunicorn==23.0.0
bcrypt==4.2.1

